Refactor image existence check
Replace the previous method of checking for the existence of a specified version with a more robust approach using the `docker manifest inspect` command. This change enhances the reliability of the check and simplifies the code. If the image exists, the script will now output '1'; otherwise, it will output '0'.

The previous code relied on extracting versions from a response variable, which is now replaced with the Docker command for manifest inspection.