Update comments in 'check_image_version.sh'
This commit updates the comments within the 'check_image_version.sh' script to improve clarity. It changes the comments to use 'true' when referring to the existence of the Docker image and 'false' when it does not exist, instead of the previous '1' and '0', making the comments more accurately describe the script's behavior.