Refactor version extraction in deployment script
This commit simplifies the version extraction process in the deployment script. Previously, the script contained a function that was only used when the script was executed directly. This function has been removed, and the version extraction logic has been moved to the top level of the script. This change makes the script easier to understand and maintain.

In addition, the deployment workflow has been updated to call the script directly, instead of piping the output of a curl command into the script. This change makes the workflow easier to read and understand.

These changes do not affect the functionality of the deployment process. The version of the code CLI is still extracted from the Visual Studio Code updates page and used to tag the Docker images.